Add setup script and instructions.
Part of the script uses a separately supplied private key to clone dependency code from a private github repo.

Later dependencies should be added to the requirements.txt added here unless we require more from anaconda, at which point an environment.yml should be added and used in the setup script. I think this is likely.