#2 (ASR): Use IBM ASR in the service using a prerecorded file
The previously explored library, SpeechRecognition, seems to not be up for the task as it doesn't handle the change APIs from cloud vendors or Python versions from the last couple of years. That lib is going out.

IBM Watson API key not committed here - solution to be made soon.
Next step, get this to use input from a microphone.