#2 (ASR): Add modified versions of files from the speech_recognition library
The lib (https://github.com/Uberi/speech_recognition) is out of date, even after the downgrade. It has classes that do convenient jobs processing microphone input though, so a modified subset is added here for now (with credit and original license). This could be replaced with a more lightweight solution, but probably won't be as I have bigger fish to fry.