#3 Add dockerfile and docker-compose (WIP)
Currently this starts but the app crashes.
The ENTRYPOINT of the dockerfile runs bash in the environment with all the dependencies, but not the commands to run python inside the .sh

docker-compose.yml exposes an arbitrary port; there is nothing listening there.

To build, run `docker-compose build` from this directory.
`docker-compose up` to run.