#1 add a working flask server for input/output
The server takes in the data that it will be sent, and it parses it to a form that (I hope) can be used by Ayesaac. 

This needs integrating with the rest of the system. 

There is some boilerplate and print statements within just to make it easy for the next person to get it to work. They can be removed later.