#3 Fix broken services: missing dependencies, typo, off by default.
Pandas is used to load a csv in colour detection, but was not installed (and requires Python 3.6.10 explicitly,  though this is what was already used); intents from upstream in NLU includes recognise, not recognize; NLG commented out.

One dependency of keras-ocr (so our OCR) was able to detect the presence of conda and do the wrong thing. Set it to install with conda to save it from itself.
There might be an override/clash with installing from multiple locations, but it won't use the one from pip if conda exists, which was the whole problem!