feat: add new service to handle end of rabbitmq pipeline
Data from the pipeline goes to this service, which then saves the data to a txt file so it can be queried by the flask server. I wasnt sure how else to do it because I dont think python can have a persistent state or something like react, and the only other way I could think of storing the data was within a database which wouldve just overcomplicated this further.